+++ foo_ui_columns/vis_spectrum_renderer.h
@@ -18,7 +18,8 @@ public:
 
     ~SpectrumAnalyserRenderer() { m_render_thread.reset(); }
 
-    void configure(Mode mode, Scale horizontal_scale, COLORREF foreground_colour, COLORREF background_colour);
+    void configure(
+        Mode mode, Scale horizontal_scale, uint32_t fft_size, COLORREF foreground_colour, COLORREF background_colour);
     void start();
     void request_stop();
     void stop();
@@ -41,6 +42,7 @@ private:
 
     Mode m_mode{};
     Scale m_horizontal_scale{};
+    uint32_t m_fft_size{default_fft_size};
     COLORREF m_foreground_colour{};
     COLORREF m_background_colour{};